refactor: reorganize service layer into domain-based packages
