+++ chaekit-spring/src/main/java/qwerty/chaekit/service/group/GroupService.java
@@ -1,4 +1,4 @@
-package qwerty.chaekit.service;
+package qwerty.chaekit.service.group;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
