+++ chaekit-spring/src/main/java/qwerty/chaekit/service/member/publisher/PublisherService.java
@@ -1,4 +1,4 @@
-package qwerty.chaekit.service;
+package qwerty.chaekit.service.member.publisher;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
