+++ chaekit-spring/src/main/java/qwerty/chaekit/service/member/user/UserService.java
@@ -1,4 +1,4 @@
-package qwerty.chaekit.service;
+package qwerty.chaekit.service.member.user;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
